GBS-23 | `SlugUtil` Class Has Been Moved to `common.util` Package
