+++ src/main/java/org/gelecekbilimde/scienceplatform/post/service/impl/PostServiceImpl.java
@@ -4,6 +4,7 @@ import lombok.RequiredArgsConstructor;
 import org.gelecekbilimde.scienceplatform.auth.model.Identity;
 import org.gelecekbilimde.scienceplatform.common.model.BasePage;
 import org.gelecekbilimde.scienceplatform.common.model.BaseSpecification;
+import org.gelecekbilimde.scienceplatform.media.util.SlugUtil;
 import org.gelecekbilimde.scienceplatform.post.model.Post;
 import org.gelecekbilimde.scienceplatform.post.model.entity.PostEntity;
 import org.gelecekbilimde.scienceplatform.post.model.enums.Process;
@@ -16,7 +17,6 @@ import org.gelecekbilimde.scienceplatform.post.repository.PostRepository;
 import org.gelecekbilimde.scienceplatform.post.service.PostMediaService;
 import org.gelecekbilimde.scienceplatform.post.service.PostProcessService;
 import org.gelecekbilimde.scienceplatform.post.service.PostService;
-import org.gelecekbilimde.scienceplatform.post.util.SlugUtil;
 import org.springframework.data.domain.Page;
 import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
